Telegram Ban Bot
Telegram bot for add and remove users from a banned list. If they choose
to respond in the group the bot replies with a message stating that they
are banned. NOTE: the bot will have access to all messages sent in the
chat and this is designed to be deployed on Google App Engine